First Part of the project is done (+-)
Meaning that we need to fit the results print in the python code.

Note that to save time and writing- all the logic has been made in c, and is written in separate code files (in the res folder)
so what remains to the second part of the project is the interface in C that is going to call that code again.